david: rutas pasadas a funciones
