+++ api/src/functions/Functions_handlers.js
@@ -1,4 +1,38 @@
-export default function pitagorasDistance(coord1, coord2){
+const { User, Review } = require("../db")
+const { updateRating } = require("../functions/Functions_user.js")
+
+//FUNCION PARA BUSCAR Y PROMEDIAR EL RATING
+const searchRating = async (id, rating) =>{
+    try {
+        const ratingValue = await User.findByPk(id,{
+            include:[
+                {
+                    model: Review,
+                    attributes: [ 'rating'],
+                    through: {attributes: []},
+                }
+            ]
+        })
+
+        if(ratingValue.rating === -1) return await updateRating(id, rating)
+        let catidadReview= ratingValue.dataValues.reviews.length
+        const reviewNumber = parseInt(catidadReview)
+
+        let ratingOld = ratingValue.dataValues.reviews.length && ratingValue.dataValues.reviews.map(obj=>obj.rating).reduce( (a,p)=> a + p, 0 )
+        const ratingTotalOld = parseInt(ratingOld)
+
+        let ratingNew = ratingTotalOld  / reviewNumber 
+        return await updateRating(id, (ratingNew + "").slice(0,3))
+        
+    } catch (error) {
+        console.log(error)
+        throw error
+    }
+}
+
+
+//FUNCION PARA SACAR LA DISTANCIA EN CORDENADAS
+const pitagorasDistance = (coord1, coord2) => {
     let x1 = coord1[0];
     let y1 = coord1[1];
 
@@ -14,6 +48,8 @@ export default function pitagorasDistance(coord1, coord2){
     return (Math.sqrt(powx+powy)*100);
 }
 
+
+
 // const closeToOne = (coords1, coords2) => {
 //     if(pitagorasDistance(coords1, coords2) < distance){ //distancia en kilometros
 //         return true;
@@ -22,3 +58,7 @@ export default function pitagorasDistance(coord1, coord2){
 //     }
 // }
 
+module.exports = {
+    searchRating,
+    pitagorasDistance,
+}
