+++ api/src/routes/jobs/jobs.route.js
@@ -1,5 +1,6 @@
 const { Router } = require('express');
 const functions = require("../../functions/Functions_user");
+const functionsJob = require("../../functions/Functions_jobs")
 const { Profession } = require("../../db.js")
 
 const jobs = Router()
@@ -7,7 +8,7 @@ const jobs = Router()
 jobs.get("/", (req, res, next) => {
     functions.getAllJobs()
     .then(jobs => {
-        return jobs ? res.send(jobs) : `No se encontraron trabajos registrados`
+        return jobs ? res.status(200).send(jobs) : `No se encontraron trabajos registrados`
     }, error => {
         return res.status(404).send(error)
     })
@@ -19,24 +20,13 @@ jobs.get("/", (req, res, next) => {
 jobs.post("/create", async (req, res, next) =>{
     const { name } = req.body;
     const jobsMinuscule = name.toLowerCase();
-    try {
-        const [newJob, created] = await Profession.findOrCreate({
-            where:{
-                name: jobsMinuscule,
-            },
-            defaults:{
-                name: jobsMinuscule,
-            }
-        })
-
-        if(!created)  res.status(200).send(`The Profession cannot be created, the Job "${jobsMinuscule}" has already exist`);
-        return res.status(201).send(`The Profession "${jobsMinuscule}" created successfully`);
-        
-    } catch (error) {
-        console.log(error)
-        next(error)
-    }
-    
+        try {
+            const newJob = await functionsJob.postJobs( jobsMinuscule )
+            res.status(201).send(newJob)
+        } catch (error) {
+            console.log(error)
+            next(error)
+        }
 })
 
 // RUTA PARA ELIMINAR TRABAJOS
