+++ api/src/routes/orders/orders.route.js
@@ -10,7 +10,6 @@ orders.get("/all", async (req, res, next)=>{
     try {
         const allOrders = await Orders.findAll()
         res.status(200).json(allOrders)
-    
     } catch (error) {
         console.log(error)
         next(error)
@@ -23,20 +22,18 @@ orders.post("/", async (req, res, next) =>{
     try {
         const newOrder = await functions.postOrder( id_user_professional, id_user_client )
         res.status(201).send(newOrder)
-
     } catch (error) {
         console.log(error)
         next(error)
     }
 })
 
-
+// RUTA PARA COMPLETAR POR EL PROFESIONAL LAS ORDENES CREADAS
 orders.put("/:id", async (req, res, next) =>{
     const { description, complete, apointment_date, allowReview, isActive } = req.body;
-    //console.log('description desde el back', description)
     const { id } = req.params;
     try {
-        functions.updateOrden(id, description, complete, apointment_date, allowReview, isActive)
+        await functions.updateOrden(id, description, complete, apointment_date, allowReview, isActive)
         res.status(200).send(`The order updated successfully`)
     } catch (error) {
         console.log(error)
